# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UpcycleMatch API",
        "description": "API Documentation untuk UpcycleMatch. Mendukung JWT Bearer Token dan Basic Auth.",
        "contact": {
            "email": "admin@upcyclematch.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/textiles": {
            "get": {
                "tags": [
                    "Textiles"
                ],
                "summary": "Daftar Semua Limbah Kain",
                "description": "Mengambil daftar limbah kain. Endpoint ini bisa diakses secara publik.",
                "operationId": "3345dea86a6bef2182399540e6d6541a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data limbah"
                    }
                }
            },
            "post": {
                "tags": [
                    "Textiles"
                ],
                "summary": "Tambah Data Limbah",
                "description": "Hanya untuk role Contributor. Menambahkan limbah baru ke sistem.",
                "operationId": "3890475d3dc3c519fd0bbf51b5917d03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "fabric_type",
                                    "weight",
                                    "address"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Kain Perca Katun"
                                    },
                                    "fabric_type": {
                                        "type": "string",
                                        "example": "Katun"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "example": 5.5
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Limbah berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Hanya Contributor yang bisa menambah"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/textiles/{id}": {
            "get": {
                "tags": [
                    "Textiles"
                ],
                "summary": "Detail Limbah Kain",
                "description": "Mengambil detail dari satu data limbah kain berdasarkan ID.",
                "operationId": "5bd8ebdcd415a0a0c60e922ccd6af0c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID limbah kain",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail limbah"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}